code climate fix;
