+++ app/javascript/Application/components/Assessment/index.js
@@ -12,3 +12,4 @@ export { default as DomainItemList } from './DomainItemList'
 export { default as DomainScore } from './DomainScore'
 export { default as DomainCaregiverControls } from './DomainCaregiverControls'
 export { default as ChangeLogPage } from './ChangeLog/ChangeLogPage'
+export { AssessmentStatus, sortAssessmentsByDate } from './AssessmentHelper'
